factor out matching to a new file
